requirements.txt
Django==2.2.5
pytz==2018.4
django-guardian==2.0.0
